cambios menores
seleccionar el gestor de cola de trabajos desde parametro en globals.h

+++ src/func.c
@@ -468,7 +468,8 @@ void marcarSigCambio(struct proceso *p)
     return;
 }
 
-float actualizarTiempoParaTerminar(int tEjec, int viejoN, int nuevoN)
+// AQUI DEBE DE ESTAR EL ERROR
+float actualizarTiempoParaTerminar(float tEjec, int viejoN, int nuevoN)
 {
     return tEjec * viejoN / nuevoN;
 }
